Add files via upload
A program which will check the users' eligibility for a scholarship, where the user will have to qualify by both gender and age. In this program, for the user to qualify for the scholarship, it is must the gender be male of age between 18 to 35. Here females of age less than 18 and more than 35 will not qualify by both gender and age, but they'll qualify only by age between 18 to 35 and still not by gender. And the males of age less than 18 and of age greater than 35 will also not qualify.